requirements.txt
pymodbus
requests
python-dotenv
paho-mqtt
influxdb-client